+++ packages/pglite/src/interface.ts
@@ -33,6 +33,7 @@ export interface ExecProtocolOptions {
   syncToFs?: boolean
   throwOnError?: boolean
   onNotice?: (notice: NoticeMessage) => void
+  dataTransferContainer?: DataTransferContainer
 }
 
 export interface ExtensionSetupResult<TNamespace = any> {
@@ -77,6 +78,8 @@ export interface DumpDataDirResult {
   filename: string
 }
 
+export type DataTransferContainer = 'cma' | 'file'
+
 export interface PGliteOptions<TExtensions extends Extensions = Extensions> {
   dataDir?: string
   username?: string
@@ -91,6 +94,7 @@ export interface PGliteOptions<TExtensions extends Extensions = Extensions> {
   fsBundle?: Blob | File
   parsers?: ParserOptions
   serializers?: SerializerOptions
+  defaultDataTransferContainer?: DataTransferContainer
 }
 
 export type PGliteInterface<T extends Extensions = Extensions> =